removes code from test file and adds it to CLI file
